+++ RazorSvelte/App/offcanvas-nav.svelte
@@ -1,0 +1,14 @@
+<script lang="ts">
+    import Layout from "./shared/layout/offcanvas-layout.svelte";
+</script>
+
+<Layout>
+    <div class="container mt-3">
+        hello off canvas
+    </div>
+    
+</Layout>
+
+<style lang="scss">
+</style>
+