+++ RazorSvelte/App/shared/urls.ts
@@ -3,6 +3,8 @@
 const urls = getFromJson<{
     authorizedUrl: string;
     indexUrl: string;
+    privacyUrl: string;
+    offcanvasNavUrl: string;
     loginUrl: string;
     logoutUrl: string;
     aboutUrl: string;