+++ RazorSvelte/Pages/OffcanvasNav.cshtml.cs
@@ -1,0 +1,3 @@
+namespace RazorSvelte.Pages;
+
+public class OffcanvasNavModel : PageModel {}