+++ RazorSvelte/Pages/OffcanvasNav.entry.ts
@@ -1,0 +1,3 @@
+/// <reference types="svelte" />
+import App from "../App/offcanvas-nav.svelte";
+export default new App({target: document.body});