+++ RazorSvelte/Pages/OffcanvasNav.rollup.config.js
@@ -1,0 +1,2 @@
+import config from "../Scripts/rollup.config";
+export default config({ input: "./Pages/OffcanvasNav.entry.ts", jsOutput: "./wwwroot/build/offcanvas-nav.js", cssOutput: "offcanvasnav.css" });